size_t -> int
git-svn-id: http://beta.visl.sdu.dk/svn/visl/tools/benchmarks@9442 cb2587b3-b7ff-0310-8c81-a2a651690ada

+++ src/string-to-int.cpp
@@ -36,7 +36,7 @@
 #include <iomanip>
 #include <sstream>
 
-static const size_t N = 100000;
+static const int N = 100000;
 static const size_t R = 7;
 
 int naive(const char *p) {
